Improve database
